refactor: oc_resolvers.py
